requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
streamlit
langchain-google-genai
Pillow
tensorflow
mysqlclient
langchain
langchain-core 
langchain-community
jwt
opencv-python
streamlit-option-menu
python-multipart